Android: Update SDL to 2.0.7, set minSDK to 14, drop MIPS, support arm64-v8a
